requirements.txt
requests==2.31.0
tweepy==4.14.0
twscrape==0.6.0
aiohttp==3.9.1
openai==1.3.0
python-dotenv==1.0.0
schedule==1.2.0